On OS upgrade, use old conffiles
When upgrading to the next debian release, we want
config file the were modified manually or touched
by the leap_platform to *not* get the newer version
the package maintainer ships in the newer version
of the package.
This is because we stepped into two issues (see below)
where the newer, package provided conffile breaks some
service.
We are aweare of possible issues that we might face where
an old config file doesn't work with a newer package, but
until we don't step into any of these, this is the better
approach.

- Resolves: #8042, #8044